requirements.txt
amqp==5.0.1
billiard==3.6.3.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==2.2.17
djangorestframework==3.12.2
kombu==5.0.2
prompt-toolkit==3.0.8
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
validate-email==1.3
vine==5.0.0
wcwidth==0.2.5
